System Tests > Refactor package structure
