+++ src/systemTest/java/com/hivemq/cli/commands/cli/publish/PublishConnectTlsST.java
@@ -16,9 +16,9 @@
 package com.hivemq.cli.commands.cli.publish;
 
 import com.google.common.io.Resources;
-import com.hivemq.cli.utils.ExecutionResultAsync;
-import com.hivemq.cli.utils.HiveMQ;
-import com.hivemq.cli.utils.MqttCliAsync;
+import com.hivemq.cli.utils.cli.results.ExecutionResultAsync;
+import com.hivemq.cli.utils.broker.HiveMQ;
+import com.hivemq.cli.utils.cli.MqttCliAsync;
 import com.hivemq.cli.utils.MqttVersionConverter;
 import org.junit.jupiter.api.Timeout;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -30,8 +30,8 @@ import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-import static com.hivemq.cli.utils.assertions.ConnectAssertion.assertConnectPacket;
-import static com.hivemq.cli.utils.assertions.PublishAssertion.assertPublishPacket;
+import static com.hivemq.cli.utils.broker.assertions.ConnectAssertion.assertConnectPacket;
+import static com.hivemq.cli.utils.broker.assertions.PublishAssertion.assertPublishPacket;
 
 public class PublishConnectTlsST {
