+++ src/systemTest/java/com/hivemq/cli/commands/cli/subscribe/SubscribeConnectTlsST.java
@@ -16,11 +16,10 @@
 package com.hivemq.cli.commands.cli.subscribe;
 
 import com.google.common.io.Resources;
-import com.hivemq.cli.utils.ExecutionResultAsync;
-import com.hivemq.cli.utils.HiveMQ;
-import com.hivemq.cli.utils.MqttCliAsync;
+import com.hivemq.cli.utils.cli.results.ExecutionResultAsync;
+import com.hivemq.cli.utils.broker.HiveMQ;
+import com.hivemq.cli.utils.cli.MqttCliAsync;
 import com.hivemq.cli.utils.MqttVersionConverter;
-import com.hivemq.extension.sdk.api.packets.general.MqttVersion;
 import com.hivemq.extension.sdk.api.packets.general.Qos;
 import com.hivemq.extension.sdk.api.packets.subscribe.RetainHandling;
 import com.hivemq.extension.sdk.api.packets.subscribe.Subscription;
@@ -34,8 +33,8 @@ import org.junit.jupiter.params.provider.ValueSource;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-import static com.hivemq.cli.utils.assertions.ConnectAssertion.assertConnectPacket;
-import static com.hivemq.cli.utils.assertions.SubscribeAssertion.assertSubscribePacket;
+import static com.hivemq.cli.utils.broker.assertions.ConnectAssertion.assertConnectPacket;
+import static com.hivemq.cli.utils.broker.assertions.SubscribeAssertion.assertSubscribePacket;
 import static org.junit.jupiter.api.Assertions.fail;
 
 public class SubscribeConnectTlsST {
