+++ src/systemTest/java/com/hivemq/cli/commands/shell/ShellDisconnectST.java
@@ -17,9 +17,9 @@
 package com.hivemq.cli.commands.shell;
 
 import com.google.common.collect.ImmutableList;
-import com.hivemq.cli.utils.AwaitOutput;
-import com.hivemq.cli.utils.HiveMQ;
-import com.hivemq.cli.utils.MqttCliShell;
+import com.hivemq.cli.utils.cli.results.AwaitOutput;
+import com.hivemq.cli.utils.broker.HiveMQ;
+import com.hivemq.cli.utils.cli.MqttCliShell;
 import com.hivemq.extensions.packets.general.UserPropertiesImpl;
 import com.hivemq.mqtt.message.mqtt5.MqttUserProperty;
 import org.jetbrains.annotations.NotNull;
@@ -33,7 +33,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-import static com.hivemq.cli.utils.assertions.DisconnectAssertion.assertDisconnectPacket;
+import static com.hivemq.cli.utils.broker.assertions.DisconnectAssertion.assertDisconnectPacket;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class ShellDisconnectST {
