+++ src/systemTest/java/com/hivemq/cli/commands/shell/ShellExitST.java
@@ -15,8 +15,8 @@
  */
 package com.hivemq.cli.commands.shell;
 
-import com.hivemq.cli.utils.HiveMQ;
-import com.hivemq.cli.utils.MqttCliShell;
+import com.hivemq.cli.utils.broker.HiveMQ;
+import com.hivemq.cli.utils.cli.MqttCliShell;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
