+++ src/systemTest/java/com/hivemq/cli/utils/cli/results/AwaitOutput.java
@@ -13,24 +13,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.cli.utils;
+package com.hivemq.cli.utils.cli.results;
 
+import com.hivemq.cli.utils.cli.io.LogWaiter;
+import com.hivemq.cli.utils.cli.io.ProcessIO;
+import com.hivemq.cli.utils.exceptions.TimeoutException;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import org.junit.jupiter.api.Assertions;
 
-import java.io.IOException;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-public class ExecutionResultAsync {
+public class AwaitOutput {
 
     private final @NotNull ProcessIO processIO;
     private final @NotNull String command;
+    private final @Nullable LogWaiter logWaiter;
 
-    public ExecutionResultAsync(final @NotNull ProcessIO processIO, final @NotNull String command) {
+    public AwaitOutput(final @NotNull ProcessIO processIO, final @Nullable LogWaiter logWaiter, final @NotNull String command) {
         this.processIO = processIO;
         this.command = command;
+        this.logWaiter = logWaiter;
     }
 
-    public @NotNull ExecutionResultAsync awaitStdOut(final @NotNull String expectedOutput) {
+    public @NotNull AwaitOutput awaitStdOut(final @NotNull String expectedOutput) {
         try {
             processIO.awaitStdOut(expectedOutput);
         } catch (final TimeoutException timeoutException) {
@@ -39,7 +45,7 @@ public class ExecutionResultAsync {
         return this;
     }
 
-    public @NotNull ExecutionResultAsync awaitStdErr(final @NotNull String expectedOutput) {
+    public @NotNull AwaitOutput awaitStdErr(final @NotNull String expectedOutput) {
         try {
             processIO.awaitStdErr(expectedOutput);
         } catch (final TimeoutException timeoutException) {
@@ -48,7 +54,14 @@ public class ExecutionResultAsync {
         return this;
     }
 
-    public void write(final @NotNull String output) throws IOException {
-        processIO.writeMsg(output);
+    public @NotNull AwaitOutput awaitLog(final @NotNull String expectedLogMessage) {
+        assertNotNull(logWaiter);
+        try {
+            logWaiter.awaitLog(expectedLogMessage);
+        } catch (final TimeoutException timeoutException) {
+            Assertions.fail(String.format("Command '%s' did not return expected logfile output '%s' in time. Actual read logfile output: '%s'", command, expectedLogMessage, timeoutException.getActualOutput()), timeoutException);
+        }
+        return this;
     }
+
 }
