+++ src/systemTest/java/com/hivemq/cli/utils/cli/results/ExecutionResult.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.cli.utils;
+package com.hivemq.cli.utils.cli.results;
 
 import org.jetbrains.annotations.NotNull;
