+++ src/systemTest/java/com/hivemq/cli/utils/cli/results/ExecutionResultAsync.java
@@ -13,27 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.cli.utils;
+package com.hivemq.cli.utils.cli.results;
 
+import com.hivemq.cli.utils.cli.io.ProcessIO;
+import com.hivemq.cli.utils.exceptions.TimeoutException;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import org.junit.jupiter.api.Assertions;
 
-import static org.junit.jupiter.api.Assertions.assertNotNull;
+import java.io.IOException;
 
-public class AwaitOutput {
+public class ExecutionResultAsync {
 
     private final @NotNull ProcessIO processIO;
     private final @NotNull String command;
-    private final @Nullable LogWaiter logWaiter;
 
-    public AwaitOutput(final @NotNull ProcessIO processIO, final @Nullable LogWaiter logWaiter, final @NotNull String command) {
+    public ExecutionResultAsync(final @NotNull ProcessIO processIO, final @NotNull String command) {
         this.processIO = processIO;
         this.command = command;
-        this.logWaiter = logWaiter;
     }
 
-    public @NotNull AwaitOutput awaitStdOut(final @NotNull String expectedOutput) {
+    public @NotNull ExecutionResultAsync awaitStdOut(final @NotNull String expectedOutput) {
         try {
             processIO.awaitStdOut(expectedOutput);
         } catch (final TimeoutException timeoutException) {
@@ -42,7 +41,7 @@ public class AwaitOutput {
         return this;
     }
 
-    public @NotNull AwaitOutput awaitStdErr(final @NotNull String expectedOutput) {
+    public @NotNull ExecutionResultAsync awaitStdErr(final @NotNull String expectedOutput) {
         try {
             processIO.awaitStdErr(expectedOutput);
         } catch (final TimeoutException timeoutException) {
@@ -51,14 +50,7 @@ public class AwaitOutput {
         return this;
     }
 
-    public @NotNull AwaitOutput awaitLog(final @NotNull String expectedLogMessage) {
-        assertNotNull(logWaiter);
-        try {
-            logWaiter.awaitLog(expectedLogMessage);
-        } catch (final TimeoutException timeoutException) {
-            Assertions.fail(String.format("Command '%s' did not return expected logfile output '%s' in time. Actual read logfile output: '%s'", command, expectedLogMessage, timeoutException.getActualOutput()), timeoutException);
-        }
-        return this;
+    public void write(final @NotNull String output) throws IOException {
+        processIO.writeMsg(output);
     }
-
 }
